Trac 2910 Apparent, while unrelated, error. Hope we haven't rounded any of genes so far...

+++ atlas-data-storage/src/main/java/uk/ac/ebi/gxa/data/NetCDFDataCreator.java
@@ -242,7 +242,7 @@ public class NetCDFDataCreator {
                 new Dimension[]{designElementDimension, designElementLenDimension}
         );
         netCdf.addVariable(
-                "GN", DataType.DOUBLE,
+                "GN", DataType.LONG,
                 new Dimension[]{designElementDimension}
         );
